requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.3
h11==0.14.0
idna==3.7
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.7.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0